# Optimize contract performance