--- a/pkg/ast/ast_operation_definition.go
+++ b/pkg/ast/ast_operation_definition.go
@@ -1,6 +1,8 @@
 package ast
 
 import (
+	"math"
+
 	"github.com/jensneuse/graphql-go-tools/internal/pkg/unsafebytes"
 	"github.com/jensneuse/graphql-go-tools/pkg/lexer/position"
 )
@@ -66,4 +68,24 @@ func (d *Document) AddVariableDefinitionToOperationDefinition(operationDefinitio
 	ref := len(d.VariableDefinitions) - 1
 	d.OperationDefinitions[operationDefinitionRef].VariableDefinitions.Refs =
 		append(d.OperationDefinitions[operationDefinitionRef].VariableDefinitions.Refs, ref)
+}
+
+const (
+	alphabet = `abcdefghijklmnopqrstuvwxyz`
+)
+
+func (d *Document) GenerateUnusedVariableDefinitionName(operationDefinition int) []byte {
+	for i := 1;i<math.MaxInt64;i++{
+		out := make([]byte,i)
+		for j := range alphabet {
+			for k := 0;k<i;k++{
+				out[k] = alphabet[j]
+			}
+			_,exists := d.VariableDefinitionByNameAndOperation(operationDefinition,out)
+			if !exists {
+				return out
+			}
+		}
+	}
+	return nil
 }