--- a/pkg/astprinter/astprinter.go
+++ b/pkg/astprinter/astprinter.go
@@ -256,7 +256,7 @@ func (p *printVisitor) EnterOperationDefinition(ref int) {
 }
 
 func (p *printVisitor) LeaveOperationDefinition(ref int) {
-	if !p.document.OperationDefinitionIsLastRootNode(ref) {
+	if !p.document.NodeIsLastRootNode(ast.Node{Kind: ast.NodeKindOperationDefinition,Ref: ref}) {
 		if p.indent != nil {
 			p.write(literal.LINETERMINATOR)
 			p.write(literal.LINETERMINATOR)