--- a/pkg/execution/datasource/datasource_http_json.go
+++ b/pkg/execution/datasource/datasource_http_json.go
@@ -28,33 +28,33 @@ var httpJsonSchemes = []string{
 // HttpJsonDataSourceConfig is the configuration object for the HttpJsonDataSource
 type HttpJsonDataSourceConfig struct {
 	// Host is the hostname of the upstream
-	Host string
+	Host string `json:"host"`
 	// URL is the url of the upstream
-	URL string
+	URL string `json:"url"`
 	// Method is the http.Method, e.g. GET, POST, UPDATE, DELETE
 	// default is GET
-	Method *string
+	Method *string `json:"method"`
 	// Body is the http body to send
 	// default is null/nil (no body)
-	Body *string
+	Body *string `json:"body"`
 	// Headers defines the header mappings
-	Headers []HttpJsonDataSourceConfigHeader
+	Headers []HttpJsonDataSourceConfigHeader `json:"headers"`
 	// DefaultTypeName is the optional variable to define a default type name for the response object
 	// This is useful in case the response might be a Union or Interface type which uses StatusCodeTypeNameMappings
-	DefaultTypeName *string
+	DefaultTypeName *string `json:"default_type_name"`
 	// StatusCodeTypeNameMappings is a slice of mappings from http.StatusCode to GraphQL TypeName
 	// This can be used when the TypeName depends on the http.StatusCode
-	StatusCodeTypeNameMappings []StatusCodeTypeNameMapping
+	StatusCodeTypeNameMappings []StatusCodeTypeNameMapping `json:"status_code_type_name_mappings"`
 }
 
 type StatusCodeTypeNameMapping struct {
-	StatusCode int
-	TypeName   string
+	StatusCode int    `json:"status_code"`
+	TypeName   string `json:"type_name"`
 }
 
 type HttpJsonDataSourceConfigHeader struct {
-	Key   string
-	Value string
+	Key   string `json:"key"`
+	Value string `json:"value"`
 }
 
 type HttpJsonDataSourcePlannerFactoryFactory struct {