--- a/pkg/execution/handler.go
+++ b/pkg/execution/handler.go
@@ -71,7 +71,7 @@ func (h *Handler) Handle(requestData, extraVariables []byte) (executor *Executor
 		err = report
 		return
 	}
-	normalizer := astnormalization.NewNormalizer(true)
+	normalizer := astnormalization.NewNormalizer(true,true)
 	normalizer.NormalizeOperation(&operationDocument, h.base.Definition, &report)
 	if report.HasErrors() {
 		err = report