--- a/pkg/graphql/request.go
+++ b/pkg/graphql/request.go
@@ -67,7 +67,7 @@ func (r *Request) Normalize(schema *Schema) (result NormalizationResult, err err
 		return normalizationResultFromReport(report)
 	}
 
-	normalizer := astnormalization.NewNormalizer(true)
+	normalizer := astnormalization.NewNormalizer(true,true)
 	normalizer.NormalizeOperation(&r.document, &schema.document, &report)
 	if report.HasErrors() {
 		return normalizationResultFromReport(report)