This is a proposal to improve the performance of the graphql package as well as reduce gc pressure.

Initial results:

```
go test -run=none -bench=BenchmarkExecutionEngine -count=1 -memprofile=mem.out && go tool pprof -alloc_space mem.out
goos: darwin
goarch: amd64
pkg: github.com/jensneuse/graphql-go-tools/pkg/graphql
BenchmarkExecutionEngine-16    	  723003	      1578 ns/op	  16.47 MB/s	    6990 B/op	      39 allocs/op
```

final results:

```
graphql % go test -run=none -bench=BenchmarkExecutionEngine -count=1 -memprofile=mem.out && go tool pprof -alloc_space mem.out
goos: darwin
goarch: amd64
pkg: github.com/jensneuse/graphql-go-tools/pkg/graphql
BenchmarkExecutionEngine-16    	 4696146	       242 ns/op	 107.40 MB/s	     582 B/op	      12 allocs/op
```

Overall a 6x improvement.

One thing to consider is I'm not 100% if there is no data race with re-using query plans across requests. I've tested this with parallel benchmarking and it worked fine. However I'd like to ask you to carefully review @pvormste @spetrunin @buger as this could potentially result in huge problems.