@buger 2020-06-05
Not sure if you should close it. Stuff still can be merged to master, just not backported to the 2.10 release.